moved common logic of routers and controller to abstract classes
