+++ mock-server/src/controllers/controller-base.ts
@@ -1,0 +1,11 @@
+import { DataContext } from "../data/data-context";
+
+export abstract class ControllerBase {
+  constructor(protected dataContext: DataContext) {}
+
+  protected wrapData(data: any) {
+    return {
+      data
+    }
+  }
+}
