--- mock-server/src/models/router-base.model.ts
@@ -1,5 +1,0 @@
-import { Router } from "express";
-
-export interface RouterBase {
-  router: Router
-}
